### 2023-09-16 #5852 @editorialbot
[PRE REVIEW]: BoboCEP: a Fault-Tolerant Complex Event Processing Engine for Edge Computing in Internet of Things
**Submitting author:** @r3w0p (<a href="http://orcid.org/0000-0001-5348-7068">Alexander Power</a>)
**Repository:** https://github.com/r3w0p/bobocep
**Branch with paper.md** (empty if default branch): joss
**Version:** 1.1.0
