### 2023-09-16 #5852 @danielskatz
👋 @r3w0p - note that your paper does not compile. Please follow the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) and note that you can click on the error above to find out more about it. In addition, you could work on the possibly missing DOI that editorialbot suggests, but note that it may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` after making changes to the .md file or when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
