### 2023-09-17 #5852 @danielskatz
@r3w0p - please add a bit to the paper about use cases: what real world research has and would this software support and enable?
