### 2023-09-17 59a92f3
use case added to joss paper
