+++ paper/paper.bib
@@ -28,6 +28,24 @@
   organization={IEEE}
 }
 
+@article{Rocha:2021,
+  title={An accessible smart home based on integrated multimodal interaction},
+  author={Rocha, Ana Patr{\'\i}cia and Ketsmur, Maksym and Almeida, Nuno and Teixeira, Ant{\'o}nio},
+  journal={Sensors},
+  volume={21},
+  number={16},
+  pages={5464},
+  year={2021},
+  publisher={MDPI}
+}
+
+@phdthesis{Scattone:2021,
+  title={Processamento de eventos complexos nativo de nuvem para cidades inteligentes},
+  author={Scattone, Fernando Freire},
+  year={2021},
+  school={Universidade de S{\~a}o Paulo}
+}
+
 @TechReport{W3C:WoT:TD,
   title="Web of Things (WoT) Thing Description",
   author={Kaebisch, Sebastian and Kamiya, Takuki and McCool, Michael and Charpenay, Victor and Kovatsch, Matthias},
