### 2023-09-18 #5852 @danielskatz
@r3w0p - thanks for the addition - please now address the missing DOIs, as discussed in https://github.com/openjournals/joss-reviews/issues/5852#issuecomment-1722314557
