+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={3},
   pages={1--62},
   year={2012},
-  publisher={ACM New York, NY, USA}
+  publisher={ACM New York, NY, USA},
+  doi={10.1145/2187671.2187677}
 }
 
 @article{Giatrakos:2020,
@@ -16,7 +17,8 @@
   volume={29},
   pages={313--352},
   year={2020},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s00778-019-00557-w}
 }
 
 @inproceedings{Power:2020,
@@ -25,7 +27,8 @@
   booktitle={2020 IEEE sixth international conference on big data computing service and applications (BigDataService)},
   pages={109--112},
   year={2020},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/BigDataService49289.2020.00024}
 }
 
 @article{Rocha:2021,
@@ -36,14 +39,16 @@
   number={16},
   pages={5464},
   year={2021},
-  publisher={MDPI}
+  publisher={MDPI},
+  doi={10.3390/s21165464}
 }
 
 @phdthesis{Scattone:2021,
   title={Processamento de eventos complexos nativo de nuvem para cidades inteligentes},
   author={Scattone, Fernando Freire},
   year={2021},
-  school={Universidade de S{\~a}o Paulo}
+  school={Universidade de S{\~a}o Paulo},
+  doi={10.11606/D.45.2021.tde-24032021-145027}
 }
 
 @TechReport{W3C:WoT:TD,
