### 2023-09-18 #5852 @r3w0p
@danielskatz Apologies for this oversight. I have added DOIs to all references, except for the online tech report which does not have one.
