### 2023-10-03 r3w0p/bobocep#7 @abhishektiwari
Unable to run the Flask 123 and dist abc examples
**Describe the bug**
Unable to run Flask example (`tests/postman/flask_123`) with Postman collection. Currently working on [review for JOSS submission](https://github.com/openjournals/joss-reviews/issues/5858)

```
WARNING: This is a development server. Do not use it in a production deployment. Use a production WSGI server instead.
 * Running on http://127.0.0.1:5000
Press CTRL+C to quit
[2023-10-02 21:57:49,813] ERROR in app: Exception on /data/int/1 [GET]
Traceback (most recent call last):
  File "/Users/x/.pyenv/versions/3.11.4/envs/bobocep/lib/python3.11/site-packages/flask/app.py", line 1455, in wsgi_app
    response = self.full_dispatch_request()
               ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/x/.pyenv/versions/3.11.4/envs/bobocep/lib/python3.11/site-packages/flask/app.py", line 869, in full_dispatch_request
    rv = self.handle_user_exception(e)
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/x/.pyenv/versions/3.11.4/envs/bobocep/lib/python3.11/site-packages/flask/app.py", line 867, in full_dispatch_request
    rv = self.dispatch_request()
         ^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/x/.pyenv/versions/3.11.4/envs/bobocep/lib/python3.11/site-packages/flask/app.py", line 852, in dispatch_request
    return self.ensure_sync(self.view_functions[rule.endpoint])(**view_args)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/x/Labs/joss/bobocep/app.py", line 27, in data_int
    engine.receiver.add_data(int(my_int))
    ^^^^^^
NameError: name 'engine' is not defined
```

**Expected behavior**
Postman collection should get successful response (`200`) from Flask server.

—OS: Mac
—Python: `3.11.4`

**Additional context**
Pip versions

```
attrs                     23.1.0
blinker                   1.6.2
bobocep                   1.1.0
click                     8.1.7
Flask                     3.0.0
itsdangerous              2.1.2
Jinja2                    3.1.2
jsonschema                4.19.0
jsonschema-specifications 2023.7.1
MarkupSafe                2.1.3
pip                       23.1.2
pycryptodome              3.18.0
referencing               0.30.2
rpds-py                   0.10.3
setuptools                65.5.0
Werkzeug                  3.0.0
```
