### 2023-10-03 r3w0p/bobocep#7 @abhishektiwari
Similar fate for the `flask_dist_abc_def ` ran with Python `3.10.12`. `bobocep` installed via `pip`. Unable to runs examples.

```
bobocep 1.1.0
Flask 3.0.0
```

Similar error as last,

For `run_dist_1.py`
```
WARNING: This is a development server. Do not use it in a production deployment. Use a production WSGI server instead.
 * Running on http://127.0.0.1:9091
Press CTRL+C to quit
[2023-10-02 22:22:36,866] ERROR in app: Exception on /data/str/a [GET]
Traceback (most recent call last):
  File "/Users/x/.pyenv/versions/3.10.12/envs/bobocep/lib/python3.10/site-packages/flask/app.py", line 1455, in wsgi_app
    response = self.full_dispatch_request()
  File "/Users/x/.pyenv/versions/3.10.12/envs/bobocep/lib/python3.10/site-packages/flask/app.py", line 869, in full_dispatch_request
    rv = self.handle_user_exception(e)
  File "/Users/x/.pyenv/versions/3.10.12/envs/bobocep/lib/python3.10/site-packages/flask/app.py", line 867, in full_dispatch_request
    rv = self.dispatch_request()
  File "/Users/x/.pyenv/versions/3.10.12/envs/bobocep/lib/python3.10/site-packages/flask/app.py", line 852, in dispatch_request
    return self.ensure_sync(self.view_functions[rule.endpoint])(**view_args)
  File "/Users/x/Labs/joss/bobocep/tests/postman/flask_dist_abc_def/run_dist_1.py", line 28, in data_str
    engine.receiver.add_data(str(my_str))
NameError: name 'engine' is not defined
```

For `run_dist_2.py`

```
[2023-10-02 22:26:16,987] ERROR in app: Exception on /data/str/d [GET]
Traceback (most recent call last):
  File "/Users/x/.pyenv/versions/3.10.12/envs/bobocep/lib/python3.10/site-packages/flask/app.py", line 1455, in wsgi_app
    response = self.full_dispatch_request()
  File "/Users/x/.pyenv/versions/3.10.12/envs/bobocep/lib/python3.10/site-packages/flask/app.py", line 869, in full_dispatch_request
    rv = self.handle_user_exception(e)
  File "/Users/x/.pyenv/versions/3.10.12/envs/bobocep/lib/python3.10/site-packages/flask/app.py", line 867, in full_dispatch_request
    rv = self.dispatch_request()
  File "/Users/x/.pyenv/versions/3.10.12/envs/bobocep/lib/python3.10/site-packages/flask/app.py", line 852, in dispatch_request
    return self.ensure_sync(self.view_functions[rule.endpoint])(**view_args)
  File "/Users/x/Labs/joss/bobocep/tests/postman/flask_dist_abc_def/run_dist_2.py", line 28, in data_str
    engine.receiver.add_data(str(my_str))
NameError: name 'engine' is not defined
```
