### 2023-10-03 r3w0p/bobocep#7 @r3w0p
Hello, thanks for raising an issue.

I have run these tests locally and can confirm that they work for me. I should provide more information for the Postman tests as to how they operate, because they are **not** to be executed in the same way as the unit tests in `tests_bobocep`.

In `postman/flask_123`, you need to directly run the `run.py` file. Failing to do this will mean that the code within `if __name__ == '__main__':` will not be executed. I suspect this is why you are getting the error `name 'engine' is not defined` because, indeed, `engine` is only defined within the aforementioned `if` statement!

- I personally use the PyCharm environment. If you are also using it, then you can right click `run.py` and press `Run 'run'`.
- Then, load up Postman and import the accompanying `.json` file, and then execute its GET requests. When I do this locally, I get, for example: `127.0.0.1 - - [03/Oct/2023 20:05:44] "GET /data/int/1 HTTP/1.1" 200 -`

Please let me know if you have tried the above. If it still doesn't work, then I will investigate further.

I will push some clearer notes on the Postman tests as soon as possible.

Alex.
