### 2023-10-03 r3w0p/bobocep#7 @abhishektiwari
Do these tests suppose to run inside the cloned repository?

I am installing `bobocep`, `flask`, and all other dependencies in an isolated virtual environment and executing the `run.py`. I believe that's how an actual user will consume `bobocep` – via a `pip` install.

No issues with importing the Postman collections.

I will give another shot and see if these runs in a cloned repository.
