### 2023-10-03 r3w0p/bobocep#7 @r3w0p
The Postman tests are for developer use only.

End users would `pip install bobocep` and would never be expected to run the Postman tests ever.

Running the tests requires the installation of requirements from both the `requirements.txt` and `requirements-dev.txt` files. If you have cloned `bobocep` locally, then you do not need to `pip install bobocep` as well.

Alex.
