### 2023-10-14 #5858 @abhishektiwari
@r3w0p please see my initial comments. I am still working to validate software functionality.

## Documentation

### Example usage:
Authors include automated tests to validate the software functionality but no examples implementing a real-world use case. Paper describes a use case — elderly resident were to fall in their home and require an ambulance – to describe how complex event is generated and event is triggered. I think this phenomenon will make a very good example for BoboCEP.

## Software paper

### Statement of need
Apart from dependability, BoboCEP enables detection of complex events at edge without need of streaming raw events to the cloud. In that sense approach is privacy protecting/preserving for use cases such as smart homes which is mentioned in following section of the paper. I will suggest calling that out in statement of need.

### State of the field
Paper is very light on state of the field. Paper points to a survey by Giatrakos, N.,et al and that's all. I would like authors to clearly describe how this software compares to other commonly-used frameworks in the field. Paper does not provide required context why streaming based CEP systems can't be deployed to edge/IoT devices/gateways.
