### 2023-10-16 1763659
state of field + privacy comments
37-49: state of field, existing CEP systems
50-58: privacy in edge computing
