+++ paper/paper.bib
@@ -1,3 +1,33 @@
+@article{Alwarafy:2020,
+  title={A survey on security and privacy issues in edge-computing-assisted internet of things},
+  author={Alwarafy, Abdulmalik and Al-Thelaya, Khaled A and Abdallah, Mohamed and Schneider, Jens and Hamdi, Mounir},
+  journal={IEEE Internet of Things Journal},
+  volume={8},
+  number={6},
+  pages={4004--4022},
+  year={2020},
+  publisher={IEEE}
+}
+
+@article{Boubeta:2021,
+  title={CEPchain: A graphical model-driven solution for integrating complex event processing and blockchain},
+  author={Boubeta-Puig, Juan and Rosa-Bilbao, Jes{\'u}s and Mendling, Jan},
+  journal={Expert Systems with Applications},
+  volume={184},
+  pages={115578},
+  year={2021},
+  publisher={Elsevier}
+}
+
+@inproceedings{Choochotkaew:2017,
+  title={EdgeCEP: Fully-distributed complex event processing on IoT edges},
+  author={Choochotkaew, Sunyanan and Yamaguchi, Hirozumi and Higashino, Teruo and Shibuya, Megumi and Hasegawa, Teruyuki},
+  booktitle={2017 13th International Conference on Distributed Computing in Sensor Systems (DCOSS)},
+  pages={121--129},
+  year={2017},
+  organization={IEEE}
+}
+
 @article{Cugola:2012,
   title={Processing flows of information: From data stream to complex event processing},
   author={Cugola, Gianpaolo and Margara, Alessandro},
@@ -8,6 +38,17 @@
   year={2012},
   publisher={ACM New York, NY, USA},
   doi={10.1145/2187671.2187677}
+}
+
+@article{Fazeldehkordi:2022,
+  title={A survey of security architectures for edge computing-based IoT},
+  author={Fazeldehkordi, Elahe and Gr{\o}nli, Tor-Morten},
+  journal={IoT},
+  volume={3},
+  number={3},
+  pages={332--365},
+  year={2022},
+  publisher={MDPI}
 }
 
 @article{Giatrakos:2020,
@@ -51,6 +92,16 @@
   doi={10.11606/D.45.2021.tde-24032021-145027}
 }
 
+@article{Semlali:2021,
+  title={SAT-CEP-monitor: An air quality monitoring software architecture combining complex event processing with satellite remote sensing},
+  author={Semlali, Badr-Eddine Boudriki and El Amrani, Chaker and Ortiz, Guadalupe and Boubeta-Puig, Juan and Garcia-de-Prado, Alfonso},
+  journal={Computers \& Electrical Engineering},
+  volume={93},
+  pages={107257},
+  year={2021},
+  publisher={Elsevier}
+}
+
 @TechReport{W3C:WoT:TD,
   title="Web of Things (WoT) Thing Description",
   author={Kaebisch, Sebastian and Kamiya, Takuki and McCool, Michael and Charpenay, Victor and Kovatsch, Matthias},
@@ -58,4 +109,15 @@
   type="W3C Recommendation",
   year="2020",
   note="https://www.w3.org/TR/2020/REC-wot-thing-description-20200409/"
+}
+
+@article{Xiao:2022,
+  title={CaFtR: A fuzzy complex event processing method},
+  author={Xiao, Fuyuan},
+  journal={International Journal of Fuzzy Systems},
+  volume={24},
+  number={2},
+  pages={1098--1111},
+  year={2022},
+  publisher={Springer}
 }
