### 2023-10-22 9bbcf21
use cases: action #1 and #2, need to do patterns
