### 2023-10-22 24f963e
assisted living use case patterns
