### 2023-10-22 #5858 @r3w0p
@abhishektiwari

Hello, I have made the following changes:

### Example Usage

In the documentation, I have introduced a new page called Use Cases which covers the Assisted Living scenario from the paper as best as I could. I could not replicate some elements of the Use Case because it is impractical (e.g., calling for an ambulance). However, I have provided working code snippets for patterns and actions. I tested the IFTTT and Twilio actions today and they both worked well.

The page can be found here:

https://github.com/r3w0p/bobocep/blob/joss/docs/use_cases.rst

### Statement of Need

In the paper, I have discussed privacy, edge computing, etc., in more detail. See paragraph at Line 50.

### State of the Field

In the paper, I have provided a more comprehensive review of existing CEP systems. See paragraph at Line 37.
