### 2023-11-16 #5858 @r3w0p
Hi @mahsan321 and @vissarion

> getting started document links are not working

I have pressed all of the links in the Getting Started page that I could see and they all work for me. Which were you having trouble with specifically?

I [haven't changed](https://github.com/r3w0p/bobocep/compare/main...joss) the Getting Started page during the review process, and so the current state of the page can be found live [here](https://bobocep.readthedocs.io/en/latest/getting_started.html).

If you are referring to the `[CM2012]` link that sends you to the bottom of the page, then that is supposed to happen. It is a link to the paper mentioned in the References section.

> just like the previous reviewer couldn't run the test examples

This was resolved in [this issue](https://github.com/r3w0p/bobocep/issues/7).
