### 2023-11-16 #5858 @r3w0p
@vissarion I have made some very minor changes to the paper just now: swapped two paragraphs around, clarified some points, simplified some explanations. It is the same content nonetheless.
