### 2023-11-18 2e04198
Added missing DOIs to BIB
