+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={6},
   pages={4004--4022},
   year={2020},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/JIOT.2020.3015432}
 }
 
 @article{Boubeta:2021,
@@ -16,7 +17,8 @@
   volume={184},
   pages={115578},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.eswa.2021.115578}
 }
 
 @inproceedings{Choochotkaew:2017,
@@ -25,7 +27,8 @@
   booktitle={2017 13th International Conference on Distributed Computing in Sensor Systems (DCOSS)},
   pages={121--129},
   year={2017},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/DCOSS.2017.14}
 }
 
 @article{Cugola:2012,
@@ -48,7 +51,8 @@
   number={3},
   pages={332--365},
   year={2022},
-  publisher={MDPI}
+  publisher={MDPI},
+  doi={10.3390/iot3030019}
 }
 
 @article{Giatrakos:2020,
@@ -99,7 +103,8 @@
   volume={93},
   pages={107257},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.compeleceng.2021.107257}
 }
 
 @TechReport{W3C:WoT:TD,
@@ -119,5 +124,6 @@
   number={2},
   pages={1098--1111},
   year={2022},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s40815-021-01118-6}
 }
