### 2023-11-20 r3w0p/bobocep#8 @vissarion
JOSS review
The reference item

Kaebisch, S., Kamiya, T., McCool, M., Charpenay, V., & Kovatsch, M. (2020). Web of things114
(WoT) thing description [W3C Recommendation]. World Wide Web Consortium (W3C)

would be useful to have a version number (or date) and a link to the webpage.
