### 2023-11-20 #5858 @vissarion
@r3w0p there is only one suggestion from my side, see https://github.com/r3w0p/bobocep/issues/8
