+++ paper/paper.bib
@@ -108,12 +108,13 @@
 }
 
 @TechReport{W3C:WoT:TD,
-  title="Web of Things (WoT) Thing Description",
+  title={Web of Things (WoT) Thing Description},
   author={Kaebisch, Sebastian and Kamiya, Takuki and McCool, Michael and Charpenay, Victor and Kovatsch, Matthias},
-  institution="World Wide Web Consortium (W3C)",
-  type="W3C Recommendation",
-  year="2020",
-  note="https://www.w3.org/TR/2020/REC-wot-thing-description-20200409/"
+  institution={World Wide Web Consortium (W3C)},
+  type={W3C Recommendation (9 April 2020)},
+  year={2020},
+  url={https://www.w3.org/TR/2020/REC-wot-thing-description-20200409/},
+  urldate={2020-04-09}
 }
 
 @article{Xiao:2022,
