### 2023-11-20 r3w0p/bobocep#8 @r3w0p
I have changed the BIB entry to:

```
@TechReport{W3C:WoT:TD,
  title={Web of Things (WoT) Thing Description},
  author={Kaebisch, Sebastian and Kamiya, Takuki and McCool, Michael and Charpenay, Victor and Kovatsch, Matthias},
  institution={World Wide Web Consortium (W3C)},
  type={W3C Recommendation (9 April 2020)},
  year={2020},
  url={https://www.w3.org/TR/2020/REC-wot-thing-description-20200409/},
  urldate={2020-04-09}
}
```

Which produces:

```
Kaebisch, S., Kamiya, T., McCool, M., Charpenay, V., & Kovatsch, M. (2020). Web of114
things (WoT) thing description [W3C Recommendation (9 April 2020)]. World Wide Web115
Consortium (W3C). https://www.w3.org/TR/2020/REC-wot-thing-description-20200409
```

I added the `url` attribute, and also included `(9 April 2020)` in `type`.

Is this okay?
