### 2023-11-20 b01143e
Update setup.cfg
