### 2023-11-20 4adc469
Update setup.cfg
