+++ setup.cfg
@@ -1,5 +1,6 @@
 [bumpversion]
 current_version = 1.1.0
+commit = True
 
 [bumpversion:file:setup.py]
 search = version="{current_version}"
