+++ tests/test_bobocep/test_dist/test_BoboDistributedTCP.py
@@ -181,96 +181,6 @@ class TestValid:
         # "Remote device" now within "SYNC Period"
         assert dist._devices[devices[1].urn].last_comms > last_comms
 
-    def test_resync_1c_1h_1u_1remote_aes(self):
-        # This test is for being in the "RESYNC Period",
-        # NOT the FORCE_RESYNC flag
-        logging.getLogger().setLevel(logging.DEBUG)
-
-        completed: List[BoboRunSerial] = [
-            tc_run_simple(
-                tc_pattern(name="pattern_completed"),
-                tc_event_simple(event_id="event_id_completed"),
-                run_id="run_id_completed",
-                phenomenon_name="phenom_name_completed",
-                block_index=1
-            ).serialize()]
-
-        halted: List[BoboRunSerial] = [
-            tc_run_simple(
-                tc_pattern(name="pattern_halted"),
-                tc_event_simple(event_id="event_id_halted"),
-                run_id="run_id_halted",
-                phenomenon_name="phenom_name_halted",
-                block_index=1
-            ).serialize()]
-
-        updated: List[BoboRunSerial] = [
-            tc_run_simple(
-                tc_pattern(name="pattern_updated"),
-                tc_event_simple(event_id="event_id_updated"),
-                run_id="run_id_updated",
-                phenomenon_name="phenom_name_updated",
-                block_index=1
-            ).serialize()]
-
-        devices = [
-            BoboDevice(
-                addr="127.0.0.1",
-                port=8082,
-                urn="urn:dist:1",
-                id_key="1111111111"),
-            BoboDevice(
-                addr="127.0.0.1",
-                port=8082,
-                urn="urn:dist:2",
-                id_key="2222222222")
-        ]
-
-        decider, dec_sub = tc_decider_sub([tc_phenomenon()])
-
-        dist = BoboDistributedTCP(
-            urn="urn:dist:1",
-            decider=decider,
-            devices=devices,
-            crypto=BoboDistributedCryptoAES("1234567890ABCDEF"),
-            period_ping=30,
-            period_resync=60,
-            flag_reset=False  # to prevent initial FORCE_RESYNC
-        )
-
-        dist_sub = StubDistributedSubscriber()
-        dist.subscribe(dist_sub)
-
-        # Sets Device 2 to within "RESYNC Period"
-        last_comms = BoboDistributedTCP._now() - 65
-        dist._devices[devices[1].urn].last_comms = last_comms
-
-        sleep(3)
-
-        t = Thread(target=tc_run_distributed_tcp, args=[dist])
-        t.start()
-
-        sleep(3)
-
-        dist.on_decider_update(
-            completed=completed,
-            halted=halted,
-            updated=updated,
-            local=True)
-
-        sleep(3)
-
-        dist.close()
-        dist.join()
-        t.join()
-        assert dist.is_closed()
-
-        sleep(3)
-
-        assert len(dist_sub.completed) == 1
-        assert len(dist_sub.halted) == 1
-        assert len(dist_sub.updated) == 1
-
     def test_flag_reset_1c_1h_1u_1remote_aes(self):
         logging.getLogger().setLevel(logging.DEBUG)
 
