+++ tests/test_bobocep/test_dist/test_BoboDistributedTCP.py
@@ -1,637 +1,0 @@
-# Copyright (c) 2019-2023 r3w0p
-# The following code can be redistributed and/or
-# modified under the terms of the MIT License.
-import logging
-from threading import Thread
-from time import sleep
-from typing import List
-
-import pytest
-
-from bobocep.cep.engine.decider.runserial import BoboRunSerial
-from bobocep.dist.crypto.aes import BoboDistributedCryptoAES
-from bobocep.dist.device import BoboDevice
-from bobocep.dist.dist import BoboDistributedError
-from bobocep.dist.tcp import BoboDistributedTCP
-from tests.test_bobocep.test_cep.test_engine.test_decider import tc_run_simple, \
-    tc_decider_sub
-from tests.test_bobocep.test_cep.test_event import tc_event_simple
-from tests.test_bobocep.test_cep.test_phenom import tc_pattern, \
-    tc_phenomenon
-from tests.test_bobocep.test_dist import StubDistributedSubscriber, \
-    tc_run_distributed_tcp
-
-
-# GitHub Actions: "OSError: [Errno 98] Address already in use"
-# Use a different port for each running dist in this file: 8080, 8081, 8082...
-
-
-class TestValid:
-
-    def test_sync_1c_1h_1u_1remote_aes(self):
-        logging.getLogger().setLevel(logging.DEBUG)
-
-        completed: List[BoboRunSerial] = [
-            tc_run_simple(
-                tc_pattern(name="pattern_completed"),
-                tc_event_simple(event_id="event_id_completed"),
-                run_id="run_id_completed",
-                phenomenon_name="phenom_name_completed",
-                block_index=1
-            ).serialize()]
-
-        halted: List[BoboRunSerial] = [
-            tc_run_simple(
-                tc_pattern(name="pattern_halted"),
-                tc_event_simple(event_id="event_id_halted"),
-                run_id="run_id_halted",
-                phenomenon_name="phenom_name_halted",
-                block_index=1
-            ).serialize()]
-
-        updated: List[BoboRunSerial] = [
-            tc_run_simple(
-                tc_pattern(name="pattern_updated"),
-                tc_event_simple(event_id="event_id_updated"),
-                run_id="run_id_updated",
-                phenomenon_name="phenom_name_updated",
-                block_index=1
-            ).serialize()]
-
-        devices = [
-            BoboDevice(
-                addr="127.0.0.1",
-                port=8080,
-                urn="urn:dist:1",
-                id_key="1111111111"),
-            BoboDevice(
-                addr="127.0.0.1",
-                port=8080,
-                urn="urn:dist:2",
-                id_key="2222222222")
-        ]
-
-        decider, dec_sub = tc_decider_sub([tc_phenomenon()])
-
-        dist = BoboDistributedTCP(
-            urn="urn:dist:1",
-            decider=decider,
-            devices=devices,
-            crypto=BoboDistributedCryptoAES("1234567890ABCDEF"),
-            period_ping=30,
-            period_resync=60,
-            flag_reset=False  # to prevent initial FORCE_RESYNC
-        )
-
-        dist_sub = StubDistributedSubscriber()
-        dist.subscribe(dist_sub)
-
-        # Sets Device 2 to within "SYNC Period"
-        last_comms = BoboDistributedTCP._now()
-        dist._devices[devices[1].urn].last_comms = last_comms
-
-        sleep(1)
-
-        t = Thread(target=tc_run_distributed_tcp, args=[dist])
-        t.start()
-
-        sleep(1)
-
-        dist.on_decider_update(
-            completed=completed,
-            halted=halted,
-            updated=updated,
-            local=True)
-
-        sleep(3)
-
-        dist.close()
-        dist.join()
-        t.join()
-        assert dist.is_closed()
-
[…]
