### 2023-11-20 #5858 @r3w0p
@vissarion

Latest release
- https://github.com/r3w0p/bobocep/releases/tag/1.1.1

Zenodo record
- https://zenodo.org/records/10160409
