### 2023-11-20 #5858 @vissarion
Still the title is "r3w0p/bobocep: 1.1.1" please replace with "BoboCEP: a Fault-Tolerant Complex Event Processing1
Engine for Edge Computing in Internet of Things"
