### 2023-11-21 #5858 @r3w0p
@vissarion The final proof looks good to me. Happy to move forward if you are!
