### 2023-11-21 #5858 @danielskatz
Congratulations to @r3w0p (Alexander Power) and co-authors on your publication!!

JOSS depends on volunteers; we couldn't do this without you!
